+++ app/models/atlas_engine/address_validation/validators/full_address/province_code_comparison.rb
@@ -8,6 +8,11 @@ module AtlasEngine
         class ProvinceCodeComparison < FieldComparisonBase
           extend T::Sig
 
+          sig { override.returns(T::Boolean) }
+          def relevant?
+            true
+          end
+
           sig { override.returns(T.nilable(Token::Sequence::Comparison)) }
           def sequence_comparison
             return @province_code_comparison if defined?(@province_code_comparison)